+++ src/main/java/com/playdata/kafka/consumer/StoryConsumer.java
@@ -2,14 +2,14 @@ package com.playdata.kafka.consumer;
 
 import com.playdata.kafka.config.TopicConfig;
 import com.playdata.kafka.dto.ArticleKafkaData;
-import com.playdata.task.TaskService;
+import com.playdata.task.service.TaskService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class SuccessStoryConsumer {
+public class StoryConsumer {
 
     private final TaskService taskService;
